Setting a persistent list of strings as root.
This changeset also adds logging infrastructure.  I'm not sure how useful
the logging is longer term, I may end up ripping it out again.

cffi apparently doesn't support comparison of struct objects (they
compare by object id, which is useless to us).  So now I wrap oids in a
wrapper that does support equality testing, and incidentally has a repr
that is useful for logging.  This wrap/unwrap is ugly, and I'd prefer
to contribute an enhancement to cffi for this if I can manage to do so.