Deal with half of the string/bytes issue.
In Python2 we treat strings as strings, and in python3 we convert between
string and utf8.  This still leaves dealing with bytes on python3 and unicode
on python2, but that is only complicated if we want to be able to use the same
pmem object store from both versions of python.