Beginnings of a garbage collector.
This only does the collection of orphaned objects.  The changeset
includes a bunch of debugging improvements, including entering
the real string values for PersistentList and str into the type
table so the gc debug prints can display them.  And it finally
fixes the refcounting for the type_table and the root object.

I'm going to need a leak detector test harness at some point.