Separate the POP API from the PObject API.
Most user programs will use only the PersistentObjectPool API, which are the
methods I had not put an underscore in front of.  But Persistent classes will
want most of those '_' prefixed methods...those constitute an API for working
with the persistent memory.  So I'm moving those to another class, and it is
the class that will be the __manager__ on Persistent objects.  POP mostly
delegates to it, but it does have some methods that the MemoryManager doesn't
need.  (I'll probably move root directly on to it as well.)