Simplify the handling of oid-as-tuple.
I'm no longer bothered by this: as I say in the comment, what really ought to be
the case is that cffi would support tuple so that we could just use tuple to
ensure we were working with the tuple form.  So instead I've made the converter
a method 'otuple', to get as close to that ideal as possible.