The GC is also only part of the POP api.
This introduces a bunch more crosstalk.  The idea is that the POP object is
using the MemroryManager like a Persistent class would, so the next step is to
take away the MM's knowledge of the POP internals, insofar as I can.

I also put the lock in around the GC, which I had forgotten earlier.  We
can't have another gc run, or application code, modifying th object
tree while we're tracing it.