Get rid of 'create' constructor keyword.
The idea here is that this gives us three ways to open a pool: specify we want
an existing file (open), specify we want to create a new pool (create), and
create-it-if-it-doesn't-exist (call the constructor directly.  Whether this is
a good API or not...we'll doubtless discuss that at some point.  I some ways
the code is a little less "clean" this way (a bit more redundancy), so I may
change my mind about this myself at some point.  The advantage is that it makes
the constructor a public API, which most python class constructors are, which
will match better with python programmer expectations.

I also realized that putting a lock around open is meaningless; there's no way
for two threads to be creating the same object at the same time, and open does
not modify the pool.