Rewrite the transaction handling machinery.
Added what I hope are comprehensive tests for the transaction machinery,
and thereby discovered and fixed lots of bugs.  The transaction is
now a singleton instance per MemoryManager, and tracks which sort
of transaction is currently open.  You can mix 'free' transactions
and 'context' transactions, but you have to respect the boundaries
of the context transaction.

Also, if a transaction is rolled back, we have to purge any oid/object
mappings that we cached.  To accomplish this, there is now a cache
object that manages the two mappings, which is referenced by both the
MemroyMangager ands its transaction object.

This changeset also adds a second type of debug logging ('trace').
We use a separate logger for this rather than adding a new level,
because we are a library and only applications can safely add levels.