Use memory logging and dump on quit.
This avoids interfering with the curses screen and minimizes as much
as possible the slowdown imposed by the logging.

Unfortunately if you use it with debug on, the final GC ends with a segfault,
which probably means there's something seriously wrong with the list code, but
I haven't tracked it down yet.